refactor(app): extract settings general section owners
